remove old code and rename
